+++ src/components/case_manager/styles/Table.module.css
@@ -14,13 +14,13 @@
 }
 
 .header {
-  background-color: #DFE2E6;
-  color: #4A5056;
+  background-color: #dfe2e6;
+  color: #4a5056;
 }
 
 .headerLabel {
   font-size: 17px;
-  color: #4A5056;
+  color: #4a5056;
 }
 
 .closeButton {
